+++ src/com/simplechat/protocol/Packet6NameChange.java
@@ -19,7 +19,7 @@ public class Packet6NameChange extends Packet {
 
 
     public Packet6NameChange(String name) {
-        byte[] name2 = name.substring(0, Math.min(name.length(), Byte.MAX_VALUE)).getBytes();
+        byte[] name2 = name.substring(0, Math.min(name.length(), Short.MAX_VALUE)).getBytes();
         this.data = new byte[name2.length + 5];
 
         int i = 0;
